core: move verkle-test to a separate file
